// packages/common/src/components/vscode/ts-type-resolver/types.ts
export interface ImportResourcePathPackage {
    kind: 'package';
    packageName: string;
    importPath?: string;
}

export interface ImportResourcePathRelative {
    kind: 'relative';
    importPath: string;
    sourcePath: string;
}

export interface ImportResourcePathRelativeInPackage {
    kind: 'relative-in-package';
    packageName: string;
    importPath: string;
    sourcePath: string;
}

export interface ImportResourceBruteForce {
    kind: 'bruteforce';
    packageName: string;
    typingFiles: any[];
}

export type ImportResourcePath =
    | ImportResourcePathPackage
    | ImportResourcePathRelative
    | ImportResourcePathRelativeInPackage
    | ImportResourceBruteForce;